# internal/delivery/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/public/camera_create": {
            "post": {
                "description": "Creates a new camera and returns its ID upon successful creation.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "Camera Creation",
                "parameters": [
                    {
                        "description": "Camera Creation",
                        "name": "camera",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CameraBase"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful creation, returning camera ID",
                        "schema": {
                            "$ref": "#/definitions/responses.CreationStringResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    }
                }
            }
        },
        "/public/camera_delete": {
            "delete": {
                "description": "Deletes an existing camera by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "Camera Deletion",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Camera ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful deletion"
                    },
                    "400": {
                        "description": "Invalid input or Camera ID not found",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    }
                }
            }
        },
        "/public/case_create": {
            "post": {
                "description": "Creates a new case with a photo (.jpeg / .jpg / .png / .svg) and case data in byte string.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "Case Creation",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Photo of the case",
                        "name": "photo",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Case data in byte string format",
                        "name": "byte_string",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful creation, returning case ID",
                        "schema": {
                            "$ref": "#/definitions/responses.CreationIntResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    }
                }
            }
        },
        "/public/manager_login": {
            "post": {
                "description": "Logs in a specialist and returns a JWT and refresh token upon successful login.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "Manager Login",
                "parameters": [
                    {
                        "description": "Manager Login",
                        "name": "specialist",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ManagerBase"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful login, returning JWT and refresh token",
                        "schema": {
                            "$ref": "#/definitions/responses.JWTRefresh"
                        }
                    },
                    "400": {
                        "description": "Invalid input or incorrect password / login",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    }
                }
            }
        },
        "/public/refresh": {
            "post": {
                "description": "Refreshes access and refresh tokens using a refresh token provided in the Authorization header.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "Refresh Tokens",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Refresh Token",
                        "name": "refresh",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful token refresh, returning new jwt and refresh token",
                        "schema": {
                            "$ref": "#/definitions/responses.JWTRefresh"
                        }
                    },
                    "400": {
                        "description": "No refresh token provided",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid or expired refresh token",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    }
                }
            }
        },
        "/public/specialist_login": {
            "post": {
                "description": "Logs in a specialist and returns a jwt and refresh token upon successful login.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "Specialist Login",
                "parameters": [
                    {
                        "description": "Specialist Login",
                        "name": "specialist",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SpecialistLogin"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful login, returning jwt and refresh token",
                        "schema": {
                            "$ref": "#/definitions/responses.JWTRefresh"
                        }
                    },
                    "400": {
                        "description": "Invalid input or incorrect password / login",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    }
                }
            }
        },
        "/public/specialist_register": {
            "post": {
                "description": "Registers a new specialist, uploads their photo, and returns a jwt and refresh token upon successful registration.\nAutomatically level=1, is_verified=false.\nLogin and password are required, along with a photo upload.\nThere are some validation on password: More than 8 symbols, contain at least one number, one uppercase and one lowercase letter.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "Specialist Registration with Photo Upload",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Login",
                        "name": "login",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Password",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Full Name",
                        "name": "fullname",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Photo Upload",
                        "name": "photo",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful registration, returning jwt and refresh token",
                        "schema": {
                            "$ref": "#/definitions/responses.JWTRefresh"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.CameraBase": {
            "type": "object",
            "required": [
                "coordinates",
                "description",
                "type"
            ],
            "properties": {
                "coordinates": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "description": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "models.ManagerBase": {
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "models.SpecialistLogin": {
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "responses.CreationIntResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "responses.CreationStringResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "responses.JWTRefresh": {
            "type": "object",
            "properties": {
                "JWT": {
                    "type": "string"
                },
                "RefreshToken": {
                    "type": "string"
                }
            }
        },
        "responses.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}